use cli library to start lspd
